requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
